Change the way tests are executed when run in --detach mode
The old behavior was to run the test script within backticks, capture it's
output and then relay the output through the socket.

The new behavior redirects stdout and stderr to the socket, then exec() the
test script.  This way, the 'prove' running on the server can see all the
output generated by the script when run with -v .  It also enables the test
counter in 'prove' to increment as the ok()s in the remote test  are executed