Fix an issue with running the right command line with detached tests
If a test script has a first line that contains anything but either whitespace
or a proper perl shebang line, then SouceHandler::Perl will report it's vote
as 0.3.  This is either because TAP::Parser::Source::shebang() incorrectly
returns whatever the first line is, even if it's not a properly formatted
shebang string, or TAP::Parser::SourceHandler::Perl::can_handle incorrectly
bails out too early with a vote of 0.3 if the 'shebang' data exists and does
not contain "perl"

SourceHandler::Perl::can_handle() returns a vote of 0.3 even for files where
the first line is "use strict".